requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pillow==10.4.0
pytesseract==0.3.13
pdf2image==1.17.0
numpy==1.26.4
pandas==2.2.2
requests==2.32.3